pepeline (ticdc): clean uselesss code
